Port integration tests to mocha
