+++ .eslintrc.mocha.yml
@@ -1,0 +1,10 @@
+env:
+  commonjs: true
+  es2021: true
+  es2022: true
+  mocha: true
+rules:
+  # Disable for convenience
+  react/display-name: off
+  # Disable for convenience
+  react/prop-types: off
