+++ packages/integration-test/.eslintrc.yml
@@ -1,3 +1,3 @@
 extends:
+  - ../../.eslintrc.mocha.yml
   - ../../.eslintrc.react.yml
-  - ../../.eslintrc.jest.yml
