--- packages/integration-test/babel.config.json
@@ -1,17 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-react",
-      {
-        "runtime": "classic"
-      }
-    ],
-    [
-      "@babel/preset-env",
-      {
-        "modules": "commonjs",
-        "targets": "defaults"
-      }
-    ]
-  ]
-}
