--- packages/integration-test/tsconfig.json
@@ -1,6 +1,0 @@
-// This configuration file is for VSCode only.
-{
-  "compilerOptions": {
-    "jsx": "react"
-  }
-}
